# docs/api-specification.yaml
openapi: 3.0.1
info:
  title: PayFlow Payment Orchestration API
  description: "Multi-gateway payment orchestration (Razorpay, Stripe, PayU, UPI)\
    \ with intelligent routing, 2-second failover, idempotency, webhook reconciliation\
    \ and a full audit trail. Amounts are integer paise."
  version: 1.0.0
servers:
- url: http://localhost
  description: Generated server url
security:
- apiKey: []
...
paths:
...
  /api/v1/webhooks/upi:
    post:
      tags:
      - Webhooks
      summary: UPI (NPCI) callback receiver (#12)
      description: "X-NPCI-Signature: base64 SHA256withRSA of the raw body"
      operationId: upi
      parameters:
      - name: X-NPCI-Signature
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              format: byte
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "409":
          description: Idempotency conflict or invalid state transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INVALID_STATE_TRANSITION:
                  value:
                    error:
                      code: INVALID_STATE_TRANSITION
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: "Invalid state transition CREATED -> REFUNDED ... Valid\
                        \ transitions from CREATED: [ABANDONED, FAILED, ROUTE_SELECTED]"
        "503":
          description: Service temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                SERVICE_UNAVAILABLE:
                  value:
                    error:
                      code: SERVICE_UNAVAILABLE
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: The service is temporarily overloaded. Please retry
                        shortly.
        "422":
          description: Business rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INVALID_CAPTURE_AMOUNT:
                  value:
                    error:
                      code: INVALID_CAPTURE_AMOUNT
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: capture amount must be between 1 and the remaining
                        hold
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                NOT_FOUND:
                  value:
                    error:
                      code: NOT_FOUND
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: payment 9f1c... not found
        "401":
          description: Missing/invalid API key or webhook signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  value:
                    error:
                      code: UNAUTHORIZED
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: A valid X-API-Key header is required.
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                VALIDATION_FAILED:
                  value:
                    error:
                      code: VALIDATION_FAILED
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: request validation failed
  /api/v1/webhooks/stripe:
    post:
      tags:
      - Webhooks
      summary: Stripe webhook receiver (#10)
      description: "Stripe-Signature: t=<unix>,v1=<hex HMAC-SHA256 of t.body>"
      operationId: stripe
      parameters:
      - name: Stripe-Signature
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              format: byte
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "409":
          description: Idempotency conflict or invalid state transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INVALID_STATE_TRANSITION:
                  value:
                    error:
                      code: INVALID_STATE_TRANSITION
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: "Invalid state transition CREATED -> REFUNDED ... Valid\
                        \ transitions from CREATED: [ABANDONED, FAILED, ROUTE_SELECTED]"
        "503":
          description: Service temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                SERVICE_UNAVAILABLE:
                  value:
                    error:
                      code: SERVICE_UNAVAILABLE
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: The service is temporarily overloaded. Please retry
                        shortly.
        "422":
          description: Business rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INVALID_CAPTURE_AMOUNT:
                  value:
                    error:
                      code: INVALID_CAPTURE_AMOUNT
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: capture amount must be between 1 and the remaining
                        hold
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                NOT_FOUND:
                  value:
                    error:
                      code: NOT_FOUND
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: payment 9f1c... not found
        "401":
          description: Missing/invalid API key or webhook signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  value:
                    error:
                      code: UNAUTHORIZED
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: A valid X-API-Key header is required.
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                VALIDATION_FAILED:
                  value:
                    error:
                      code: VALIDATION_FAILED
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: request validation failed
  /api/v1/webhooks/razorpay:
    post:
      tags:
      - Webhooks
      summary: Razorpay webhook receiver (#9)
      description: "X-Razorpay-Signature: hex HMAC-SHA256 of the raw body"
      operationId: razorpay
      parameters:
      - name: X-Razorpay-Signature
        in: header
        description: hex HMAC-SHA256
        required: false
        schema:
          type: string
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              format: byte
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "409":
          description: Idempotency conflict or invalid state transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INVALID_STATE_TRANSITION:
                  value:
                    error:
                      code: INVALID_STATE_TRANSITION
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: "Invalid state transition CREATED -> REFUNDED ... Valid\
                        \ transitions from CREATED: [ABANDONED, FAILED, ROUTE_SELECTED]"
        "503":
          description: Service temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                SERVICE_UNAVAILABLE:
                  value:
                    error:
                      code: SERVICE_UNAVAILABLE
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: The service is temporarily overloaded. Please retry
                        shortly.
        "422":
          description: Business rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INVALID_CAPTURE_AMOUNT:
                  value:
                    error:
                      code: INVALID_CAPTURE_AMOUNT
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: capture amount must be between 1 and the remaining
                        hold
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                NOT_FOUND:
                  value:
                    error:
                      code: NOT_FOUND
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: payment 9f1c... not found
        "401":
          description: Missing/invalid API key or webhook signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  value:
                    error:
                      code: UNAUTHORIZED
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: A valid X-API-Key header is required.
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                VALIDATION_FAILED:
                  value:
                    error:
                      code: VALIDATION_FAILED
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: request validation failed
  /api/v1/webhooks/payu:
    post:
      tags:
      - Webhooks
      summary: PayU webhook receiver (#11)
      description: "X-PayU-Signature: hex HMAC-SHA512 of the raw body"
      operationId: payu
      parameters:
      - name: X-PayU-Signature
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              format: byte
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
        "409":
          description: Idempotency conflict or invalid state transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INVALID_STATE_TRANSITION:
                  value:
                    error:
                      code: INVALID_STATE_TRANSITION
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: "Invalid state transition CREATED -> REFUNDED ... Valid\
                        \ transitions from CREATED: [ABANDONED, FAILED, ROUTE_SELECTED]"
        "503":
          description: Service temporarily unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                SERVICE_UNAVAILABLE:
                  value:
                    error:
                      code: SERVICE_UNAVAILABLE
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: The service is temporarily overloaded. Please retry
                        shortly.
        "422":
          description: Business rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INVALID_CAPTURE_AMOUNT:
                  value:
                    error:
                      code: INVALID_CAPTURE_AMOUNT
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: capture amount must be between 1 and the remaining
                        hold
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                NOT_FOUND:
                  value:
                    error:
                      code: NOT_FOUND
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: payment 9f1c... not found
        "401":
          description: Missing/invalid API key or webhook signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  value:
                    error:
                      code: UNAUTHORIZED
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: A valid X-API-Key header is required.
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                VALIDATION_FAILED:
                  value:
                    error:
                      code: VALIDATION_FAILED
                      details: {}
                      timestamp: 2025-03-19T10:30:00.000Z
                      request_id: req_a1b2c3d4e5f6
                      message: request validation failed
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: PAYMENT_AUTH_FAILED
            message:
              type: string
              example: Payment authorisation was declined by the issuing bank.
            details:
              type: object
              example: "{gateway_error_code=BAD_REQUEST_ERROR, suggestion=Please try\
                \ a different payment method or contact your bank., gateway=razorpay,\
                \ gateway_error_description=The card issuer has declined this transaction.}"
            request_id:
              type: string
              example: req_a1b2c3d4e5f6
            trace_id:
              type: string
              example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
            timestamp:
              type: string
              example: 2025-03-19T10:30:00.000Z
      description: A7.2 standard error format
  securitySchemes:
    apiKey:
      type: apiKey
      name: X-API-Key
      in: header